+++ foundation/org.eclipse.persistence.core/src/org/eclipse/persistence/oxm/record/JsonGeneratorRecord.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013, 2015  Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2018  Oracle and/or its affiliates. All rights reserved.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 and Eclipse Distribution License v. 1.0
  * which accompanies this distribution.
@@ -112,7 +112,7 @@ public class JsonGeneratorRecord extends JsonRecord<JsonRecord.Level> {
             if((schemaType == null || theClass == null) && (CoreClassConstants.NUMBER.isAssignableFrom(value.getClass()))){
                 //if it's still a number and falls through the cracks we dont want "" around the value
                     BigDecimal convertedNumberValue = ((BigDecimal) ((ConversionManager) session.getDatasourcePlatform().getConversionManager()).convertObject(value, CoreClassConstants.BIGDECIMAL, schemaType));
-                    jsonGenerator.write(keyName, (BigDecimal)convertedNumberValue);
+                    jsonGenerator.write(keyName, convertedNumberValue);
             }else{
                 jsonGenerator.write(keyName, convertedValue);
             }
@@ -148,7 +148,7 @@ public class JsonGeneratorRecord extends JsonRecord<JsonRecord.Level> {
             if((schemaType == null || theClass == null) && (CoreClassConstants.NUMBER.isAssignableFrom(value.getClass()))){
                 //if it's still a number and falls through the cracks we dont want "" around the value
                     BigDecimal convertedNumberValue = ((BigDecimal) ((ConversionManager) session.getDatasourcePlatform().getConversionManager()).convertObject(value, CoreClassConstants.BIGDECIMAL, schemaType));
-                    jsonGenerator.write((BigDecimal)convertedNumberValue);
+                    jsonGenerator.write(convertedNumberValue);
 
             }else{
                 jsonGenerator.write(convertedValue);
